// src/models/IoK8sApiPolicyV1PodDisruptionBudgetList/types.ts
import { IoK8sApiPolicyV1PodDisruptionBudgetModelOptions as IoK8sApiPolicyV1PodDisruptionBudget } from '../IoK8sApiPolicyV1PodDisruptionBudget';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiPolicyV1PodDisruptionBudgetListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiPolicyV1PodDisruptionBudget>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}